+++ kupyna/src/lib.rs
@@ -24,18 +24,10 @@ digest::buffer_ct_variable!(
     pub struct KupynaShort<OutSize>(block_api::KupynaShortVarCore);
     max_size: U32;
 );
-digest::buffer_rt_variable!(
-    /// Short Kupyna variant which allows to select output size at runtime.
-    pub struct KupynaShortVar(block_api::KupynaShortVarCore);
-);
 digest::buffer_ct_variable!(
     /// Long Kupyna variant generic over output size.
     pub struct KupynaLong<OutSize>(block_api::KupynaLongVarCore);
     max_size: U64;
-);
-digest::buffer_rt_variable!(
-    /// Long Kupyna variant which allows to select output size at runtime.
-    pub struct KupynaLongVar(block_api::KupynaLongVarCore);
 );
 
 /// Kupyna-224 hasher.